requirements.txt
fpdf2
pandas
openai
statsbombpy